+++ spec/rates_store/memory_spec.rb
@@ -15,7 +15,7 @@ describe Money::RatesStore::Memory do
     end
 
     context ':without_mutex' do
-      let(:subject) { Money::RatesStore::Memory.new(:without_mutex => true) }
+      let(:subject) { Money::RatesStore::Memory.new(without_mutex: true) }
 
       it "doesn't use mutex if requested not to" do
         expect(subject.instance_variable_get(:@mutex)).not_to receive(:synchronize)
@@ -58,12 +58,23 @@ describe Money::RatesStore::Memory do
     end
 
     context 'no mutex' do
-      let(:subject) { Money::RatesStore::Memory.new(:without_mutex => true) }
+      let(:subject) { Money::RatesStore::Memory.new(without_mutex: true) }
 
       it 'does not use mutex' do
         expect(subject.instance_variable_get('@mutex')).not_to receive(:synchronize)
         subject.transaction{ 1 + 1 }
       end
+    end
+  end
+
+  describe '#marshal_dump' do
+    let(:subject) { Money::RatesStore::Memory.new(optional: true) }
+
+    it 'can reload' do
+      bank = Money::Bank::VariableExchange.new(subject)
+      bank = Marshal.load(Marshal.dump(bank))
+      expect(bank.store.instance_variable_get(:@options)).to eq subject.instance_variable_get(:@options)
+      expect(bank.store.instance_variable_get(:@index)).to eq subject.instance_variable_get(:@index)
     end
   end
 end